requirements.txt
amqp==2.3.2
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-activity-stream==0.7.0
django-autofixture==0.12.1
django-cors-headers==2.5.2
django-countries==5.3.3
django-datatables-view==1.17.0
django-js-asset==1.2.2
django-jsonfield==1.1.0
django-jsonfield-compat==0.4.4
django-mptt==0.9.1
django-qr-code==1.0.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-simple-history==2.7.0
djangorestframework==3.9.2
et-xmlfile==1.0.1
honcho==1.0.1
idna==2.8
itypes==1.1.0
jdcal==1.4
Jinja2==2.10.1
kombu==4.2.2.post1
MarkupSafe==1.1.1
openapi-codec==1.3.2
openpyxl==2.6.2
Pillow==5.4.0
psycopg2==2.8
ptvsd==4.2.8
Pygments==2.3.1
pytz==2018.7
qrcode==6.1
reportlab==3.5.18
requests==2.21.0
simple-history==1.1
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.1
validate-email==1.3
vine==1.1.4
